[ANDROSDK-900] Move metadata download to module and rename
