+++ core/src/androidTest/java/org/hisp/dhis/android/core/dataset/internal/DataSetCompleteRegistrationPostCallRealIntegrationShould.java
@@ -64,7 +64,7 @@ public class DataSetCompleteRegistrationPostCallRealIntegrationShould extends Ba
     //@Test
     public void upload_data_set_complete_registrations_with_to_post_state() throws Exception {
         d2.userModule().logIn("android", "Android123").blockingGet();
-        d2.syncMetaData().blockingSubscribe();
+        d2.metadataModule().blockingDownload();
         d2.aggregatedModule().data().download().subscribe();
 
         DataSetCompleteRegistration dataSetCompleteRegistration
@@ -87,7 +87,7 @@ public class DataSetCompleteRegistrationPostCallRealIntegrationShould extends Ba
     //@Test
     public void upload_data_set_complete_registrations_with_to_update_state() throws Exception {
         d2.userModule().logIn("android", "Android123").blockingGet();
-        d2.syncMetaData().blockingSubscribe();
+        d2.metadataModule().blockingDownload();
         d2.aggregatedModule().data().download().subscribe();
 
         DataSetCompleteRegistration dataSetCompleteRegistration
@@ -107,7 +107,7 @@ public class DataSetCompleteRegistrationPostCallRealIntegrationShould extends Ba
     //@Test
     public void update_and_delete_different_data_set_complete_registrations() throws Exception {
         d2.userModule().logIn("android", "Android123").blockingGet();
-        d2.syncMetaData().blockingSubscribe();
+        d2.metadataModule().blockingDownload();
         d2.aggregatedModule().data().download().subscribe();
 
         DataSetCompleteRegistration toDeleteDataSetCompleteRegistration
@@ -133,7 +133,7 @@ public class DataSetCompleteRegistrationPostCallRealIntegrationShould extends Ba
     //@Test
     public void delete_data_set_complete_registrations_with_to_delete_state() throws Exception {
         d2.userModule().logIn("android", "Android123").blockingGet();
-        d2.syncMetaData().blockingSubscribe();
+        d2.metadataModule().blockingDownload();
         d2.aggregatedModule().data().download().subscribe();
 
         DataSetCompleteRegistration dataSetCompleteRegistration
