+++ core/src/androidTest/java/org/hisp/dhis/android/core/enrollment/note/NotePostCallRealIntegrationShould.java
@@ -77,7 +77,7 @@ public class NotePostCallRealIntegrationShould extends BaseRealIntegrationTest {
     public void downloadUpdateAndSyncTei() throws Exception {
         d2.userModule().logIn(RealServerMother.user, RealServerMother.password).blockingGet();
 
-        d2.syncMetaData().blockingSubscribe();
+        d2.metadataModule().blockingDownload();
 
         d2.trackedEntityModule().downloadTrackedEntityInstancesByUid(Lists.newArrayList("AlvUHPP2Mes")).blockingGet();
