+++ core/src/androidTest/java/org/hisp/dhis/android/core/trackedentity/search/TrackedEntityInstanceQueryAndDownloadRealIntegrationShould.java
@@ -68,7 +68,7 @@ public class TrackedEntityInstanceQueryAndDownloadRealIntegrationShould extends 
     public void query_and_download_tracked_entity_instances() throws Exception {
         login();
 
-        d2.syncMetaData().blockingSubscribe();
+        d2.metadataModule().blockingDownload();
 
         TrackedEntityInstanceQuery query = queryBuilder.build();
         List<TrackedEntityInstance> queriedTeis =
