+++ core/src/androidTest/java/org/hisp/dhis/android/core/utils/integration/mock/BaseMockIntegrationTestMetadataEnqueable.java
@@ -37,7 +37,7 @@ public abstract class BaseMockIntegrationTestMetadataEnqueable extends BaseMockI
         boolean isNewInstance = setUpClass(MockIntegrationTestDatabaseContent.MetadataEnqueable);
         if (isNewInstance) {
             objects.dhis2MockServer.enqueueMetadataResponses();
-            objects.d2.syncMetaData().blockingSubscribe();
+            objects.d2.metadataModule().blockingDownload();
         }
     }
 }
