+++ core/src/androidTest/java/org/hisp/dhis/android/core/wipe/WipeDBCallRealIntegrationShould.java
@@ -60,7 +60,7 @@ public class WipeDBCallRealIntegrationShould extends BaseRealIntegrationTest {
     public void have_empty_database_when_wipe_db_after_sync_metadata() throws Exception {
         d2.userModule().logIn(RealServerMother.user, RealServerMother.password).blockingGet();
 
-        d2.syncMetaData().blockingSubscribe();
+        d2.metadataModule().blockingDownload();
 
         DatabaseAssert.assertThatDatabase(databaseAdapter()).isNotEmpty();
 
@@ -73,7 +73,7 @@ public class WipeDBCallRealIntegrationShould extends BaseRealIntegrationTest {
     public void have_empty_database_when_wipe_db_after_sync_data() throws Exception {
         d2.userModule().logIn(RealServerMother.user, RealServerMother.password).blockingGet();
 
-        d2.syncMetaData().blockingSubscribe();
+        d2.metadataModule().blockingDownload();
 
         Callable<List<Event>> eventCall = EventCallFactory.create(d2.retrofit(), d2.databaseAdapter(), "DiszpKrYNg8", 0);
 
@@ -90,7 +90,7 @@ public class WipeDBCallRealIntegrationShould extends BaseRealIntegrationTest {
     public void do_not_have_metadata_when_wipe_metadata_after_sync_metadata() throws Exception {
         d2.userModule().logIn(RealServerMother.user, RealServerMother.password).blockingGet();
 
-        d2.syncMetaData().blockingSubscribe();
+        d2.metadataModule().blockingDownload();
 
         DatabaseAssert.assertThatDatabase(databaseAdapter()).isNotEmpty();
 
@@ -103,7 +103,7 @@ public class WipeDBCallRealIntegrationShould extends BaseRealIntegrationTest {
     public void do_not_have_data_when_wipe_data_after_sync() throws Exception {
         d2.userModule().logIn(RealServerMother.user, RealServerMother.password).blockingGet();
 
-        d2.syncMetaData().blockingSubscribe();
+        d2.metadataModule().blockingDownload();
 
         d2.trackedEntityModule().downloadTrackedEntityInstances(5, false, false).subscribe();
