+++ core/src/main/java/org/hisp/dhis/android/core/domain/metadata/MetadataCall.java
@@ -141,4 +141,8 @@ public class MetadataCall {
 
                 })), true);
     }
+
+    public void blockingDownload() {
+        download().blockingSubscribe();
+    }
 }
